# api.yml
openapi: 3.0.3
info:
  title: RSU/ESOP Platform API
  description: |
    API for managing equity compensation (RSUs/stock options) for startups.
    Features include pool management, grant administration, vesting calculations, and employee portal access.
    
    Key platform features:
    - Single ordinary-share class
    - Fixed DECIMAL(12,3) precision for all share & monetary amounts
    - 12-month cliff, 48-month total vesting
    - Roles: Admin, Employee
    - Pool & PPS management with vesting engine
    - Daily vesting batch job
    - Soft delete on all tables
  version: '1.0.0'
  contact:
    name: Platform Support
    email: support@rsu-platform.com
servers:
  - url: 'https://api.rsu-platform.com/api'
    description: Production API server
  - url: 'https://staging.rsu-platform.com/api'
    description: Staging server for testing
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0-issued JWT Access Token.
        Obtained via Auth0 Universal Login flow.
        Contains standard claims plus potential custom claims for role/permissions.
        Example claims expected by backend middleware:
        - sub: Auth0 User ID (used as auth0_user_id)
        - aud: API Identifier (audience)
        - iss: Issuer URL
        - exp: Expiration timestamp
        - iat: Issued at timestamp
        - scope: Permissions (if using RBAC)
        - Custom claims for role/tenant_id (if configured in Auth0 Rules/Actions)
  schemas:
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              example: 'INVALID_GRANT'
            message:
              type: string
              description: Human-readable error message
              example: 'The specified grant does not exist or is inaccessible'
            details:
              type: object
              description: Field-specific validation errors
              example:
                grant_id: 'Invalid UUID format'
                share_amount: 'Must be greater than 0.001'
                status: 'Must be one of: active, inactive'
    Success:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data, schema varies by endpoint
...
    Grant:
      type: object
      required:
        - grant_id
        - tenant_id
        - employee_id
        - grant_date
        - share_amount
        - vested_amount
        - status
        - created_at
      properties:
        grant_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        tenant_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        employee_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        grant_date:
          type: string
          format: date
          example: '2025-05-01'
        share_amount:
          $ref: '#/components/schemas/DecimalString'
          description: 'Total shares granted (must be > 0)'
        vested_amount:
          $ref: '#/components/schemas/DecimalString'
          description: 'Default is 0.000, increases as shares vest. Atomically synchronized with VestingEvents.'
        status:
          type: string
          enum: [active, inactive]
          example: 'active'
          description: 'Changes from active to inactive upon termination'
        termination_date:
          type: string
          format: date
          nullable: true
          example: null
          description: 'Date of grant termination, if applicable'
        unvested_shares_returned:
          $ref: '#/components/schemas/DecimalString'
          description: 'Calculated as share_amount - vested_amount at termination'
        termination_reason:
          type: string
          nullable: true
          example: null
          description: 'Reason for termination, if applicable'
        terminated_by:
          type: string
          format: uuid
          nullable: true
          example: null
          description: 'User who performed the termination, if applicable'
        version:
          type: integer
          minimum: 0
          example: 0
          description: 'Incremented on each update for optimistic locking'
        created_by:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
          description: 'User who created the grant'
        created_at:
          type: string
          format: date-time
          example: '2025-05-01T10:00:00Z'
        deleted_at:
          type: string
          format: date-time
          nullable: true
          example: null
          description: 'Soft delete timestamp, if applicable'
    GrantCreate:
      type: object
      required:
        - employee_id
        - grant_date
        - share_amount
      properties:
        employee_id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        grant_date:
          type: string
          format: date
          example: '2025-05-01'
        share_amount:
          $ref: '#/components/schemas/DecimalString'
          description: 'Must be at least 0.001 and not exceed pool Available amount'
...
    PaginatedResponse:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            type: object
        pagination:
          type: object
          required:
            - total_items
            - total_pages
            - current_page
            - limit
          properties:
            total_items:
              type: integer
              example: 45
            total_pages:
              type: integer
              example: 3
            current_page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            next_page:
              type: integer
              nullable: true
              example: 2
            prev_page:
              type: integer
              nullable: true
              example: null
...
    DecimalString:
      type: string
      description: String representation of a DECIMAL(12,3) value.
      pattern: '^-?[0-9]{1,9}(\\.[0-9]{1,3})?$'
      example: '123456.789'
...
  parameters:
...
    GrantIdParam:
      name: grant_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the grant
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page for pagination
    StatusParam:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [active, inactive]
      description: Filter by status
    FromDateParam:
      name: from_date
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Filter by date range (start date)
    ToDateParam:
      name: to_date
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Filter by date range (end date)
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
paths:
...
  /grants:
    post:
      summary: Create grant
      description: Create a new grant for an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantCreate'
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Grant created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Grant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
    
    get:
      summary: List grants
      description: Retrieve paginated list of grants
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StatusParam'
        - name: employee_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter by employee ID
        - $ref: '#/components/parameters/FromDateParam'
        - $ref: '#/components/parameters/ToDateParam'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Grants retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PaginatedResponse'
                          - type: object
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Grant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /grants/{grant_id}:
    get:
      summary: Get grant details
      description: Retrieve specific grant details
      parameters:
        - $ref: '#/components/parameters/GrantIdParam'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Grant details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Grant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    
    patch:
      summary: Update grant
      description: Update grant details (limited to non-financial fields)
      parameters:
        - $ref: '#/components/parameters/GrantIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version:
                  type: integer
                  example: 0
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Grant updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Grant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict - Version mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
...